refactor(app): Overhaul component lifecycle and reload logic
This commit introduces a major architectural refactoring to a more robust, component-based design, improving maintainability, scalability, and reliability.

The previous architecture centralized component management and task handling within the `Application` class. This made the class complex and the configuration reload logic fragile and difficult to maintain.

Key changes include:

- **Component Protocol**: A new `Component` protocol with `start()` and `stop()` methods is introduced to create a unified interface for all major services (`Scanner`, `MqttClient`, `Exporter`, `AutomationHandler`).

- **Decentralized Lifecycle Management**: Each component is now responsible for managing its own lifecycle, including its own `asyncio.Task`. The `Application` class is simplified and no longer manages a central list of tasks.

- **Robust Reloading (SIGHUP)**: The configuration reload mechanism has been completely rewritten. It now follows a safer "stop-all, start-all" pattern. This includes robust error handling to roll back to the previous stable configuration if the new configuration fails to start.

- **Test Overhaul**: The test suite, particularly `test_app.py`, has been rewritten to reflect the new architecture. `test_exporter.py` was also improved to use an isolated registry, preventing test pollution.

+++ switchbot_actions/config.py
@@ -33,7 +33,7 @@ class MqttSettings(BaseConfigModel):
     port: int = Field(1883, ge=1, le=65535)
     username: Optional[str] = None
     password: Optional[str] = None
-    reconnect_interval: int = 10
+    reconnect_interval: float = 10.0
 
 
 class PrometheusExporterSettings(BaseConfigModel):
